feat: add image validation for trustyai operator and service
