+++ tests/model_explainability/conftest.py
@@ -2,8 +2,12 @@ from typing import Generator, Any
 
 import pytest
 from kubernetes.dynamic import DynamicClient
+from ocp_resources.config_map import ConfigMap
 from ocp_resources.namespace import Namespace
 from ocp_resources.persistent_volume_claim import PersistentVolumeClaim
+from pytest_testconfig import config as py_config
+
+from tests.model_explainability.trustyai_service.trustyai_service_utils import TRUSTYAI_SERVICE_NAME
 
 
 @pytest.fixture(scope="class")
@@ -19,3 +23,15 @@ def pvc_minio_namespace(
         size="10Gi",
     ) as pvc:
         yield pvc
+
+
+@pytest.fixture(scope="session")
+def trustyai_operator_configmap(
+    admin_client: DynamicClient,
+) -> ConfigMap:
+    return ConfigMap(
+        client=admin_client,
+        namespace=py_config["applications_namespace"],
+        name=f"{TRUSTYAI_SERVICE_NAME}-operator-config",
+        ensure_exists=True,
+    )
