--- tests/model_explainability/trustyai_operator/conftest.py
@@ -1,19 +1,0 @@
-import pytest
-from kubernetes.dynamic import DynamicClient
-from ocp_resources.config_map import ConfigMap
-
-from tests.model_explainability.trustyai_service.trustyai_service_utils import TRUSTYAI_SERVICE_NAME
-
-from pytest_testconfig import config as py_config
-
-
-@pytest.fixture(scope="session")
-def trustyai_operator_configmap(
-    admin_client: DynamicClient,
-) -> ConfigMap:
-    return ConfigMap(
-        client=admin_client,
-        namespace=py_config["applications_namespace"],
-        name=f"{TRUSTYAI_SERVICE_NAME}-operator-config",
-        ensure_exists=True,
-    )
